+++ src/alg.rs
@@ -97,7 +97,7 @@ impl<T: fmt::Debug + 'static> std::error::Error for WeakKeyError<T> {}
 /// # use rand::thread_rng;
 /// # use jwt_compact::{prelude::*, alg::{Hs256, Hs256Key, StrongAlg, StrongKey}};
 /// # fn main() -> anyhow::Result<()> {
-/// let weak_key = Hs256Key::from(b"too short!" as &[u8]);
+/// let weak_key = Hs256Key::new(b"too short!");
 /// assert!(StrongKey::try_from(weak_key).is_err());
 /// // There is no way to create a `StrongKey` from `weak_key`!
 ///