+++ tests/algorithms.rs
@@ -35,18 +35,14 @@ fn hs256_reference() {
     const KEY: &str =
         "AyM1SysPpbyDfgZld3umj1qzKObwVMkoqQ-EstJQLr_T-1qS0gZH75aKtMN3Yj0iPS4hcgUuTwjAzZr1Z9CAow";
 
-    let token = UntrustedToken::try_from(TOKEN).unwrap();
+    let token = UntrustedToken::new(TOKEN).unwrap();
     assert_eq!(token.algorithm(), "HS256");
 
     let key = base64::decode_config(KEY, base64::URL_SAFE_NO_PAD).unwrap();
-    let key = Hs256Key::from(key.as_slice());
+    let key = Hs256Key::new(&key);
     let validated_token = Hs256.validate_integrity::<Obj>(&token, &key).unwrap();
     assert_eq!(
-        validated_token
-            .claims()
-            .expiration_date
-            .unwrap()
-            .timestamp(),
+        validated_token.claims().expiration.unwrap().timestamp(),
         1_300_819_380
     );
     assert_eq!(validated_token.claims().custom["iss"], json!("joe"));
@@ -79,9 +75,9 @@ fn hs384_reference() {
          lLufmCVZRUuyTwJF311JHuh";
     const KEY: &[u8] = b"your-384-bit-secret";
 
-    let token = UntrustedToken::try_from(TOKEN).unwrap();
+    let token = UntrustedToken::new(TOKEN).unwrap();
     assert_eq!(token.algorithm(), "HS384");
-    assert_eq!(token.header().signature_type, Some("JWT".to_owned()));
+    assert_eq!(token.header().token_type, Some("JWT".to_owned()));
 
     let key = Hs384Key::from(KEY);
     let token = Hs384
@@ -108,9 +104,9 @@ fn hs512_reference() {
          ugapLalKKDo6qAJkBy0i8d9DFcYIySIUgQ69Dprvp4fpA";
     const KEY: &[u8] = b"your-512-bit-secret";
 
-    let token = UntrustedToken::try_from(TOKEN).unwrap();
+    let token = UntrustedToken::new(TOKEN).unwrap();
     assert_eq!(token.algorithm(), "HS512");
-    assert_eq!(token.header().signature_type, Some("JWT".to_owned()));
+    assert_eq!(token.header().token_type, Some("JWT".to_owned()));
 
     let key = Hs512Key::from(KEY);
     let token = Hs512
@@ -148,7 +144,7 @@ fn es256k_reference() {
 
     let public_key = PublicKey::from_slice(&KEY).unwrap();
     let es256k: Es256k = Default::default();
-    let token = UntrustedToken::try_from(TOKEN).unwrap();
+    let token = UntrustedToken::new(TOKEN).unwrap();
     assert_eq!(token.algorithm(), "ES256K");
 
     let token = es256k
@@ -197,7 +193,7 @@ fn test_ed25519_reference() {
     check_key_traits::<EdSigningKey, EdVerifyingKey>();
 
     let public_key = EdVerifyingKey::from_slice(&KEY).unwrap();
-    let token = UntrustedToken::try_from(TOKEN).unwrap();
+    let token = UntrustedToken::new(TOKEN).unwrap();
     assert_eq!(token.algorithm(), "Ed25519");
 
     let token = Ed25519::with_specific_name()
@@ -329,7 +325,7 @@ fn create_claims() -> Claims<CompactClaims> {
 
     let mut claims = Claims::new(CompactClaims { subject: [1; 32] });
     claims.issued_at = Some(now);
-    claims.expiration_date = Some(now + Duration::days(7));
+    claims.expiration = Some(now + Duration::days(7));
     claims
 }
 
@@ -365,7 +361,7 @@ fn compact_token_hs256() {
         long_token_str.len(),
         token_str.len(),
     );
-    let untrusted_token = UntrustedToken::try_from(&*token_str).unwrap();
+    let untrusted_token = UntrustedToken::new(&token_str).unwrap();
     let token = Hs256.validate_integrity(&untrusted_token, &key).unwrap();
     assert_eq!(*token.claims(), claims);